requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.32
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
protobuf==3.20.3
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
SQLAlchemy==2.0.2
typing-extensions==4.4.0
Werkzeug==2.2.2
zipp==3.12.1
